Show questions

+++ mobile/src/main/java/onl/deepspace/zoorallye/questions/CustomSeekBar.java
@@ -15,9 +15,9 @@ import android.widget.SeekBar;
 // TODO-dennis: 01.04.2016 Add a label pin to CustomSeekBar
 public class CustomSeekBar extends SeekBar {
 
-    private float mMin = Float.POSITIVE_INFINITY;
-    private float mMax = Float.POSITIVE_INFINITY;
-    private float mStepSize = Float.POSITIVE_INFINITY;
+    private double mMin = Float.POSITIVE_INFINITY;
+    private double mMax = Float.POSITIVE_INFINITY;
+    private double mStepSize = Float.POSITIVE_INFINITY;
 
     public CustomSeekBar(Context context) {
         super(context);
@@ -42,7 +42,7 @@ public class CustomSeekBar extends SeekBar {
         super.onDraw(canvas);
     }
 
-    public void setFloatRange(float min, float max, float stepSize) {
+    public void setFloatRange(double min, double max, double stepSize) {
         mMin = min;
         mMax = max;
         mStepSize = stepSize;
@@ -50,27 +50,27 @@ public class CustomSeekBar extends SeekBar {
     }
 
     @SuppressWarnings("unused")
-    public float getFloatMin() {
+    public double getFloatMin() {
         return mMin;
     }
 
     @SuppressWarnings("unused")
-    public float getFloatMax() {
+    public double getFloatMax() {
         return mMax;
     }
 
     @SuppressWarnings("unused")
-    public float getFloatStepSize() {
+    public double getFloatStepSize() {
         return mStepSize;
     }
 
     @SuppressWarnings("unused")
-    public void setFloatProgress(float progress) {
+    public void setFloatProgress(double progress) {
         // Check if progress is between min and max;
         if(progress < mMin || progress > mMax)
             throw new IllegalStateException("progress has to be between min and max");
         // Check if you can with an natural number from
-        float delta = progress - mMin;
+        double delta = progress - mMin;
         if(isReachable(delta, mStepSize))
             throw new IllegalStateException("Make sure you can get with an natural number from min to progress");
 
@@ -78,14 +78,14 @@ public class CustomSeekBar extends SeekBar {
         setProgress(steps);
     }
 
-    public float getFloatProgress() {
+    public double getFloatProgress() {
         int normalProgress = getProgress();
         int decimals = getDecimals(mStepSize);
         return round(mMin + mStepSize * normalProgress, decimals);
     }
 
     private void calculateDimensions() {
-        float delta = mMax - mMin;
+        double delta = mMax - mMin;
 
         // Check if max is greater than min
         if(delta == 0) throw new IllegalStateException("Min and max are equals. Max has to be greater than min");
@@ -101,21 +101,21 @@ public class CustomSeekBar extends SeekBar {
         setMax(mSteps);
     }
 
-    private int getDecimals(float number) {
-        String text = Float.toString(number);
+    private int getDecimals(double number) {
+        String text = Double.toString(number);
         int integerPlaces = text.indexOf('.');
         return text.length() - integerPlaces - 1;
     }
 
-    private float round(float number, int decimalPlaces) {
-        float roundFactor = (float) Math.pow(10, decimalPlaces);
+    private double round(double number, int decimalPlaces) {
+        double roundFactor = (double) Math.pow(10, decimalPlaces);
         return Math.round(number * roundFactor) / roundFactor;
     }
 
-    private boolean isReachable(float delta, float stepSize) {
+    private boolean isReachable(double delta, double stepSize) {
         boolean isReachable = false;
 
-        float i = 0;
+        double i = 0;
         while (i < delta) {
             if(i == delta) isReachable = true;
             i += stepSize;
